Add McpRequest interface
- Add McpRequest interface
- Remove McpConnection
  - Most methods added to McpRequest instead
- Move ImplementationInfo into its own file
