+++ mcp-server-api/src/main/java/org/mcp_java/server/completion/CompleteResourceTemplate.java
@@ -22,7 +22,7 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 import org.mcp_java.server.Cancellation;
-import org.mcp_java.server.McpConnection;
+import org.mcp_java.server.McpRequest;
 import org.mcp_java.server.progress.Progress;
 import org.mcp_java.server.resources.ResourceTemplate;
 
@@ -40,7 +40,7 @@ import org.mcp_java.server.resources.ResourceTemplate;
  * with the {@link CompleteArg#name() name} attribute set, unless the code was compiled with the
  * {@code -parameters} option. The name must match the name of one of the arguments of the related resource template.
  * <li>{@link CompletionContext} - to access values already chosen for other arguments
- * <li>{@link McpConnection} - to access information about the MCP connection
+ * <li>{@link McpRequest} - to access information about the request
  * <li>{@link Cancellation} - to allow processing to be stopped if the client cancels the tool call
  * <li>{@link Progress} - to send progress reports back to the client
  * <li>Implementations may define additional types that can be used as parameters
