+++ mcp-server-api/src/main/java/org/mcp_java/server/package-info.java
@@ -24,8 +24,7 @@
  * <li>{@link org.mcp_java.server.ContentEncoder} - Custom content encoding</li>
  * <li>{@link org.mcp_java.server.Icons @Icons} - Assigns icons to tools, resources and prompts</li>
  * <li>{@link org.mcp_java.server.McpException} - Base MCP exception</li>
- * <li>{@link org.mcp_java.server.McpConnection} - Connection metadata and status, including client capabilities</li>
- * <li>{@link org.mcp_java.server.McpLog} - Client-side logging</li>
+ * <li>{@link org.mcp_java.server.McpRequest} - Request metadata</li>
  * </ul>
  */
 package org.mcp_java.server;
