+++ mcp-server-api/src/main/java/org/mcp_java/server/resources/Resource.java
@@ -22,7 +22,7 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 import org.mcp_java.server.Cancellation;
-import org.mcp_java.server.McpConnection;
+import org.mcp_java.server.McpRequest;
 import org.mcp_java.server.Role;
 import org.mcp_java.server.progress.Progress;
 
@@ -33,7 +33,7 @@ import org.mcp_java.server.progress.Progress;
  * <p>
  * Resource methods may have parameters of the following types:
  * <ul>
- * <li>{@link McpConnection} - to access information about the MCP connection
+ * <li>{@link McpRequest} - to access information about the request
  * <li>{@link Cancellation} - to allow processing to be stopped if the client cancels the resource fetch request
  * <li>{@link Progress} - to send progress reports back to the client
  * <li>Implementations may define additional types that can be used as parameters
