+++ mcp-server-api/src/main/java/org/mcp_java/server/resources/ResourceTemplate.java
@@ -22,7 +22,7 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 import org.mcp_java.server.Cancellation;
-import org.mcp_java.server.McpConnection;
+import org.mcp_java.server.McpRequest;
 import org.mcp_java.server.Role;
 import org.mcp_java.server.progress.Progress;
 
@@ -37,7 +37,7 @@ import org.mcp_java.server.progress.Progress;
  * In most cases these parameters must be annotated with{@link ResourceTemplateArg} with the {@link ResourceTemplateArg#name() name}
  * attribute set, unless the code was compiled with the {@code -parameters} option.
  * {@code ResourceTemplateArg} also allows other properties of the argument to be configured.
- * <li>{@link McpConnection} - to access information about the MCP connection
+ * <li>{@link McpRequest} - to access information about the request
  * <li>{@link Cancellation} - to allow processing to be stopped if the client cancels the tool call
  * <li>{@link Progress} - to send progress reports back to the client
  * <li>Implementations may define additional types that can be used as parameters
