+++ mcp-server-api/src/main/javadoc/overview.html
@@ -28,9 +28,7 @@ limitations under the License.
         <li>{@link org.mcp_java.server.Cancellation} - Request cancellation handling</li>
         <li>{@link org.mcp_java.server.ContentEncoder} - Custom content encoding</li>
         <li>{@link org.mcp_java.server.McpException} - Base MCP exception</li>
-        <li>{@link org.mcp_java.server.McpConnection} - Connection metadata and status, including client capabilities
-        </li>
-        <li>{@link org.mcp_java.server.McpLog} - Client-side logging</li>
+        <li>{@link org.mcp_java.server.McpRequest} - Request metadata</li>
         <li>{@link org.mcp_java.server.progress.Progress} - Progress reporting for long-running
             operations</li>
     </ul>
